% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hist.R
\name{hist.data.frame}
\alias{hist.data.frame}
\title{Overrides base \code{hist} to work on dataframes.
Iterates interactively, plots histograms & barcharts of all columns in batch.}
\usage{
\method{hist}{data.frame}(df, mfrows = 4, mfcols = 4, ...)
}
\description{
Overrides base \code{hist} to work on dataframes.
Iterates interactively, plots histograms & barcharts of all columns in batch.
}
\examples{
{
 data(mtcars)
 hist(mtcars)
 hist(iris)
}
}

